requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
MarkupSafe==2.1.1
